+++ backend/integration_test/oidc_test.go
@@ -38,9 +38,11 @@ func oidcBaseConfig(t *testing.T) config.StashSphereServeConfig {
 		},
 		Domains: struct {
 			AllowedDomains []string `koanf:"allowed"`
+			CookieDomain   string   `koanf:"cookieDomain"`
 			ApiDomain      string   `koanf:"api"`
 		}{
 			AllowedDomains: []string{"http://localhost"},
+			CookieDomain:   "",
 			ApiDomain:      "",
 		},
 		Auth: struct {